+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * You can now explicitly point items to use other items in the Override compendium. Click the D&D Beyond button on an item in the characters inventory and select the item from the drop down. @alexdm#3414
 * On the Forge default folder paths will now be set to your asset library.
 * On the Forge changes to how data is loaded meant that icon dictionaries would not load.
+* Resetting Token images did not always work.
 
 # 3.1.6